requirements.txt
python-telegram-bot==13.15
Pillow
python-dotenv